Extracts files
